requirements.txt
# ==============================================================================
# CareerAI Backend Dependencies
# ==============================================================================

# ===================
# Django Core
# ===================
Django>=5.0,<7.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=24.0

python-dotenv

# ===================
# Database
# ===================
psycopg2-binary>=2.9
dj-database-url>=2.1

# ===================
# Configuration
# ===================
python-decouple>=3.8

# ===================
# File Processing
# ===================
pillow>=10.2
PyPDF2>=3.0
python-docx>=1.0
pdfplumber>=0.10

# ===================
# AI/ML Services
# ===================
# Google Gemini (via LangChain)
langchain-google-genai

# Fallback AI provider (optional)
openai>=1.12

# ML Libraries
numpy>=1.26
scikit-learn>=1.4

# ===================
# External APIs
# ===================
requests>=2.31

# ===================
# API Documentation
# ===================
drf-spectacular>=0.27

# ===================
# Production Server
# ===================
gunicorn>=22.0
whitenoise>=6.0
uvicorn>=0.30

# ===================
# Code Quality
# ===================
black>=24.0
isort>=5.13
flake8>=7.0
mypy>=1.8

# ===================
# Utilities
# ===================
python-dateutil>=2.9
pytz>=2024.1
striprtf>=0.0.26
